requirements.txt
streamlit>=1.33,<2
PyYAML>=6,<7
